draings

+++ data_source/data_reader.py
@@ -32,7 +32,7 @@ def register_csv_dialect():
 
 class DataReader(object):
     def __init__(self):
-        self._data = np.array([], dtype=np.float32)
+        self._data = np.array([], dtype=np.float64)
         self._start_time = None
 
         PREPROCESSED_FOLDER_PATH = './data/preprocessed'
@@ -57,9 +57,9 @@ class DataReader(object):
             def reduce_px():
                 for idx in range(data_len):
                     yield float(data[idx]['Close'])
-            col_time = np.fromiter(reduce_time(), dtype=np.float32)
-            col_vol = np.fromiter(reduce_volume(), dtype=np.float32)
-            col_px = np.fromiter(reduce_px(), dtype=np.float32)
+            col_time = np.fromiter(reduce_time(), dtype=np.float64)
+            col_vol = np.fromiter(reduce_volume(), dtype=np.float64)
+            col_px = np.fromiter(reduce_px(), dtype=np.float64)
 
         else:
             files = fnmatch.filter(os.listdir('./data'), '*.csv')
@@ -70,7 +70,7 @@ class DataReader(object):
                 if file_name.startswith(get_config().ticker):
                     sdt = file_name[len(get_config().ticker) + 1:].split('.')[0]
                     start_time = datetime.datetime.strptime(sdt, '%Y%m%d_%H%M%S').replace(tzinfo=datetime.timezone.utc)
-                    raw_data = np.genfromtxt(file_path, delimiter=',', dtype=np.float32)
+                    raw_data = np.genfromtxt(file_path, delimiter=',', dtype=np.float64)
                     break
 
             if raw_data is None:
@@ -91,7 +91,7 @@ class DataReader(object):
                     yield time.timestamp()
                     time += datetime.timedelta(minutes=1)
 
-            col_raw_time = np.fromiter(time_generator(), dtype=np.float32, count=raw_data_len)
+            col_raw_time = np.fromiter(time_generator(), dtype=np.float64, count=raw_data_len)
 
             def reduce_time(minutes: int):
                 for idx in range(raw_data_len):
@@ -100,44 +100,44 @@ class DataReader(object):
                         yield t
 
             def reduce_close(minutes: int):
-                close = np.float32('nan')
+                close = np.float64('nan')
                 for idx in range(raw_data_len):
                     c = col_raw_close[idx]
                     if not np.isnan(c):
                         close = c
                     if (idx + 1) % minutes == 0:
                         yield close
-                        close = np.float32('nan')
+                        close = np.float64('nan')
 
             def reduce_high(minutes: int):
-                high = np.float32('nan')
+                high = np.float64('nan')
                 for idx in range(raw_data_len):
                     h = col_raw_high[idx]
                     if np.isnan(high) or high < h:
                         high = h
                     if (idx + 1) % minutes == 0:
                         yield high
-                        high = np.float32('nan')
+                        high = np.float64('nan')
 
             def reduce_low(minutes: int):
-                low = np.float32('nan')
+                low = np.float64('nan')
                 for idx in range(raw_data_len):
                     l = col_raw_low[idx]
                     if np.isnan(low) or low > l:
                         low = l
                     if (idx + 1) % minutes == 0:
                         yield low
-                        low = np.float32('nan')
+                        low = np.float64('nan')
 
             def reduce_open(minutes: int):
-                open = np.float32('nan')
+                open = np.float64('nan')
                 for idx in range(col_raw_volume.shape[0]):
                     o = col_raw_open[idx]
                     if np.isnan(open):
                         open = o
                     if (idx + 1) % minutes == 0:
                         yield open
-                        open = np.float32('nan')
+                        open = np.float64('nan')
 
             def reduce_volume(minutes: int):
                 volume = 0.
@@ -149,25 +149,25 @@ class DataReader(object):
                         volume = 0.
 
             def reduce_last_close(minutes: int):
-                last_close = np.float32('nan')
+                last_close = np.float64('nan')
                 for idx in range(col_raw_volume.shape[0]):
                     lc = col_raw_last_close[idx]
                     if np.isnan(last_close):
                         last_close = lc
                     if (idx + 1) % minutes == 0:
                         yield last_close
-                        last_close = np.float32('nan')
+                        last_close = np.float64('nan')
 
             data_len = raw_data_len // get_config().bar_min
 
-            col_time = np.fromiter(reduce_time(get_config().bar_min), dtype=np.float32, count=data_len)
-            col_open = np.fromiter(reduce_open(get_config().bar_min), dtype=np.float32, count=data_len)
-            col_close = np.fromiter(reduce_close(get_config().bar_min), dtype=np.float32, count=data_len)
-            col_high = np.fromiter(reduce_high(get_config().bar_min), dtype=np.float32, count=data_len)
-            col_low = np.fromiter(reduce_low(get_config().bar_min), dtype=np.float32, count=data_len)
-            col_volume = np.fromiter(reduce_volume(get_config().bar_min), dtype=np.float32, count=data_len)
-            col_last_close = np.fromiter(reduce_last_close(get_config().bar_min), dtype=np.float32, count=data_len)
-            col_ones = np.ones(col_volume.shape, dtype=np.float32)
+            col_time = np.fromiter(reduce_time(get_config().bar_min), dtype=np.float64, count=data_len)
+            col_open = np.fromiter(reduce_open(get_config().bar_min), dtype=np.float64, count=data_len)
+            col_close = np.fromiter(reduce_close(get_config().bar_min), dtype=np.float64, count=data_len)
+            col_high = np.fromiter(reduce_high(get_config().bar_min), dtype=np.float64, count=data_len)
+            col_low = np.fromiter(reduce_low(get_config().bar_min), dtype=np.float64, count=data_len)
+            col_volume = np.fromiter(reduce_volume(get_config().bar_min), dtype=np.float64, count=data_len)
+            col_last_close = np.fromiter(reduce_last_close(get_config().bar_min), dtype=np.float64, count=data_len)
+            col_ones = np.ones(col_volume.shape, dtype=np.float64)
 
             def reduce_px():
                 for idx in range(col_volume.shape[0]):
@@ -178,7 +178,7 @@ class DataReader(object):
                     else:
                         yield lc
 
-            col_px = np.fromiter(reduce_px(), dtype=np.float32, count=data_len)
+            col_px = np.fromiter(reduce_px(), dtype=np.float64, count=data_len)
             col_vol = col_volume
             # possible normalization
             # vol_mean = np.mean(col_volume)
@@ -218,7 +218,7 @@ class DataReader(object):
 
     def _restore_preprocessed_data(self):
         print('Restoring preprocessed data...')
-        self._data = np.genfromtxt(self._DATA_FILE_PATH, delimiter=',', dtype=np.float32)
+        self._data = np.genfromtxt(self._DATA_FILE_PATH, delimiter=',', dtype=np.float64)
 
     def read_data(self) -> None:
